refactor: profile info
